+++ src/main/java/com/braintreegateway/TransactionPayPalRequest.java
@@ -4,6 +4,8 @@ public class TransactionPayPalRequest extends Request {
     private TransactionRequest parent;
     private String payeeId;
     private String payeeEmail;
+    private String payerId;
+    private String paymentId;
 
     public TransactionPayPalRequest(TransactionRequest parent) {
         this.parent = parent;
@@ -19,6 +21,16 @@ public class TransactionPayPalRequest extends Request {
         return this;
     }
 
+    public TransactionPayPalRequest payerId(String payerId) {
+        this.payerId = payerId;
+        return this;
+    }
+
+    public TransactionPayPalRequest paymentId(String paymentId) {
+        this.paymentId = paymentId;
+        return this;
+    }
+
     public TransactionRequest done() {
         return parent;
     }
@@ -29,6 +41,14 @@ public class TransactionPayPalRequest extends Request {
 
     public String getPayeeEmail() {
         return payeeEmail;
+    }
+
+    public String getPayerId() {
+        return payerId;
+    }
+
+    public String getPaymentId() {
+        return paymentId;
     }
 
     @Override
@@ -49,6 +69,8 @@ public class TransactionPayPalRequest extends Request {
     protected RequestBuilder buildRequest(String root) {
         return new RequestBuilder(root).
             addElement("payeeId", payeeId).
-            addElement("payeeEmail", payeeEmail);
+            addElement("payeeEmail", payeeEmail).
+            addElement("payerId", payerId).
+            addElement("paymentId", paymentId);
     }
 }